+++ src/components/RTransitionComponent/type.ts
@@ -1,0 +1,7 @@
+import type { BaseTransitionProps } from 'vue'
+
+export interface TransitionProps {
+  transitionPropName?: string
+  transitionMode?: BaseTransitionProps['mode']
+  transitionAppear?: boolean
+}