+++ src/types/modules/element.ts
@@ -1,8 +1,0 @@
-import type { VueInstance } from './vue'
-
-export type MaybeElement =
-  | HTMLElement
-  | SVGElement
-  | VueInstance
-  | undefined
-  | null