+++ src/types/modules/utils.ts
@@ -1,5 +1,6 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
 import type CryptoJS from 'crypto-js'
+import type { BasicTarget } from './vue'
 
 export type StorageLike = 'sessionStorage' | 'localStorage'
 
@@ -56,3 +57,7 @@ export type ElementSelector = string | `attr:${string}`
 export type MaybeArray<T> = T | T[]
 
 export type DownloadAnyFileDataType = Blob | File | string | ArrayBuffer
+
+export type EventListenerTarget = BasicTarget<
+  HTMLElement | Element | Window | Document
+>